public APIs for GPU are generated, cpu and gpu functions are separated into mods, cuda tests are fixed (now panics if cuda not found)
