+++ src/lib.rs
@@ -6,42 +6,11 @@
 #[cfg(test)]
 mod tests;
 
-use std::error::Error;
-use std::fmt;
+mod cpu_sloth;
+mod gpu_sloth;
 
-extern "C" {
-    fn sloth256_189_encode(inout: *mut u8, len: usize, iv_: *const u8, layers: usize) -> bool;
-    fn sloth256_189_decode(inout: *mut u8, len: usize, iv_: *const u8, layers: usize);
-}
+use cpu_sloth::decode;
+use cpu_sloth::encode;
 
-/// Data bigger than the prime, this is not supported
-#[derive(Debug, Copy, Clone)]
-pub struct DataBiggerThanPrime;
-
-impl fmt::Display for DataBiggerThanPrime {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Data bigger than the prime, this is not supported")
-    }
-}
-
-impl Error for DataBiggerThanPrime {}
-
-/// Sequentially encodes a 4096 byte piece s.t. a minimum amount of
-/// wall clock time elapses
-pub fn encode(
-    piece: &mut [u8; 4096],
-    iv: [u8; 32],
-    layers: usize,
-) -> Result<(), DataBiggerThanPrime> {
-    unsafe {
-        if sloth256_189_encode(piece.as_mut_ptr(), piece.len(), iv.as_ptr(), layers) {
-            return Err(DataBiggerThanPrime);
-        }
-    };
-    Ok(())
-}
-
-/// Sequentially decodes a 4096 byte encoding in time << encode time
-pub fn decode(piece: &mut [u8; 4096], iv: [u8; 32], layers: usize) {
-    unsafe { sloth256_189_decode(piece.as_mut_ptr(), piece.len(), iv.as_ptr(), layers) };
-}
+use gpu_sloth::check_cuda;
+use gpu_sloth::gpu_encode;
